+++ routes/login.js
@@ -1,8 +1,12 @@
-
-/*
- * GET add page
+/**
+ * Javascript file for dealing with the login page
  */
 
-exports.view = function(req, res){
+/**
+ * Render method for the login page (aka studuwu.herokuapp.com)
+ * @param req request data going in. unused
+ * @param res response data going back out
+ */
+exports.render = function(req, res){
   res.render('login');
 };
