+++ routes/privacy.js
@@ -1,3 +1,12 @@
-exports.view = function(req, res) {
+/**
+ * Javascript file for dealing with the privacy page
+ */
+
+/**
+ * This function renders the privacy page
+ * @param req information coming in. Unused
+ * @param res information going out
+ */
+exports.render = function(req, res) {
   res.render("privacy");
 }
