+++ routes/terms.js
@@ -1,3 +1,12 @@
-exports.view = function(req, res) {
+/**
+ * Javscript file for handling the terms and conditions page
+ */
+
+/**
+ * This function renders the terms page
+ * @param req information coming in. Unused
+ * @param res information going out
+ */
+exports.render = function(req, res) {
   res.render('terms');
 }
