+++ routes/todo.js
@@ -1,4 +1,12 @@
+/**
+ * Javascript file for dealing with the todo page
+ */
 
-exports.view = function(request, response) {
+/**
+ * This function renders the todo page
+ * @param request information coming in. Unused
+ * @param response information going out
+ */
+exports.render = function(request, response) {
   response.render('toduwu');
 };
