+++ views/edit.handlebars
@@ -33,7 +33,7 @@
 </div>
 
 <!-- Settings-->
-<form role="form" action="/add/{{userID}}/subs/" method="post">
+<form role="form" action="/add/{{userID}}/send/" method="post">
 
 	<div class="settings">
