+++ views/help.handlebars
@@ -101,20 +101,7 @@
 	<h4>
 		<div class="info">
 			<a onclick="show(5)">
-			5) [Insert Question]
-			</a>
-		</div>
-		<div class="hiddenInfo" id="hidden5" style="display: none;">
-			<p>
-				- [Insert Answer]
-			</p>
-		</div>
-	</h4>
-
-	<h4>
-		<div class="info">
-			<a onclick="show(6)">
-			6) Why is there an UwU at the bottom of each page?
+			5) Why is there an UwU at the bottom of each page?
 			</a>
 		</div>
 		<div class="hiddenInfo" id="hidden6" style="display: none;">
@@ -126,8 +113,8 @@
 
 	<h4>
 		<div class="info">
-			<a onclick="show(7)">
-			7) Why Studuwu?
+			<a onclick="show(6)">
+			6) Why Studuwu?
 			</a>
 		</div>
 		<div class="hiddenInfo" id="hidden7" style="display: none;">
